requirements.txt
aiodns==4.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncio==4.0.0
attrs==25.4.0
bcrypt==5.0.0
certifi @ file:///C:/Users/task_176765924055489/croot/certifi_1767659363027/work/certifi
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dataclasses==0.6
docopt==0.6.2
dotenv==0.9.9
frozenlist==1.8.0
greenlet==3.3.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
logging==0.4.9.6
markdown-it-py==4.0.0
mdurl==0.1.2
msgspec==0.20.0
multidict==6.7.1
mysql-connector-python==9.6.0
numpy==2.4.3
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049099114/work
pandas==3.0.1
pillow==12.1.1
pipreqs==0.4.13
pluggy==1.6.0
propcache==0.4.1
pyarrow==23.0.1
pycares==5.0.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyMySQL==1.1.2
pyotp==2.9.0
PySide6==6.10.2
PySide6_Addons==6.10.2
PySide6_Essentials==6.10.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
qrcode==8.2
reportlab==4.4.10
requests==2.32.5
rich==14.3.3
setuptools==80.10.2
shellingham==1.5.4
shiboken6==6.10.2
six==1.17.0
SQLAlchemy==2.0.48
structlog==25.5.0
tenacity==9.1.4
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wheel==0.46.3
yarg==0.1.10
yarl==1.23.0
